+++ packages/blitz-auth/src/server/adapters/index.ts
@@ -1,1 +1,2 @@
 export * from "./passport/adapter"
+export * from "./next-auth/webpack"