+++ packages/blitz-auth/src/shared/types.ts
@@ -82,7 +82,7 @@ declare module "blitz" {
   export interface Ctx {
     session: SessionContext
   }
-  export interface AuthenticatedMiddlewareCtx extends Omit<Ctx, "session"> {
+  export interface AuthenticatedCtx extends Omit<Ctx, "session"> {
     session: AuthenticatedSessionContext
   }
 }