+++ apps/next13/CHANGELOG.md
@@ -1,5 +1,80 @@
 # next-blitz-auth
 
+## 0.1.1-beta.9
+
+### Patch Changes
+
+- Updated dependencies [19898a488]
+- Updated dependencies [6811eab1a]
+- Updated dependencies [022392c12]
+  - @blitzjs/rpc@2.0.0-beta.33
+  - blitz@2.0.0-beta.33
+  - @blitzjs/next@2.0.0-beta.33
+  - @blitzjs/auth@2.0.0-beta.33
+  - @blitzjs/config@2.0.0-beta.33
+
+## 0.1.1-beta.8
+
+### Patch Changes
+
+- Updated dependencies [82649f341]
+- Updated dependencies [8b01175b4]
+  - blitz@2.0.0-beta.32
+  - @blitzjs/next@2.0.0-beta.32
+  - @blitzjs/auth@2.0.0-beta.32
+  - @blitzjs/rpc@2.0.0-beta.32
+  - @blitzjs/config@2.0.0-beta.32
+
+## 0.1.1-beta.7
+
+### Patch Changes
+
+- Updated dependencies [90f1741da]
+- Updated dependencies [df3265b85]
+  - @blitzjs/auth@2.0.0-beta.31
+  - blitz@2.0.0-beta.31
+  - @blitzjs/rpc@2.0.0-beta.31
+  - @blitzjs/next@2.0.0-beta.31
+  - @blitzjs/config@2.0.0-beta.31
+
+## 0.1.1-beta.6
+
+### Patch Changes
+
+- Updated dependencies [c5572bec6]
+- Updated dependencies [727734955]
+  - @blitzjs/auth@2.0.0-beta.30
+  - blitz@2.0.0-beta.30
+  - @blitzjs/rpc@2.0.0-beta.30
+  - @blitzjs/next@2.0.0-beta.30
+  - @blitzjs/config@2.0.0-beta.30
+
+## 0.1.1-beta.5
+
+### Patch Changes
+
+- Updated dependencies [b6b9a1c5a]
+- Updated dependencies [61888d1a3]
+  - @blitzjs/auth@2.0.0-beta.29
+  - blitz@2.0.0-beta.29
+  - @blitzjs/rpc@2.0.0-beta.29
+  - @blitzjs/next@2.0.0-beta.29
+  - @blitzjs/config@2.0.0-beta.29
+
+## 0.1.1-beta.4
+
+### Patch Changes
+
+- Updated dependencies [5166e5e03]
+- Updated dependencies [2533caf48]
+- Updated dependencies [c7ac86b85]
+- Updated dependencies [1bb3a6556]
+  - @blitzjs/auth@2.0.0-beta.28
+  - @blitzjs/next@2.0.0-beta.28
+  - @blitzjs/rpc@2.0.0-beta.28
+  - blitz@2.0.0-beta.28
+  - @blitzjs/config@2.0.0-beta.28
+
 ## 0.1.1-beta.3
 
 ### Patch Changes